split Executor class into separate file
note that the Parser class is contained in lowercase parser.ts
